[None][perf] idea_0092_v1_perf_v2: hand raw bf16 compressor rows to the DSV4 sparse-attention op, convert in the current-token store (idea idea_0092)
The DSV4 compressor projections (main + indexer wkv/wgate) upcast their
bf16 GEMM outputs with a per-layer `.float()` whose only decode-window
consumer is the fused current-token store into the fp32 compressor
caches -- the compressed-row update reads the caches, not the raw rows.
Drop the modeling-side `.float()` (24 copy kernels/step: 4 ratio-4
layers x 4 + 4 ratio-128 layers x 2), let `_multi_current_token_store_kernel`
load native-dtype sources and convert on store (bf16 -> fp32 widening is
exact), and anchor the compressed-row reconstruction dtype on the fp32
cache instead of the incoming rows. The prefill/mixed branch and the
source-op reference widen once at entry, and the standalone
compressor/indexer forwards widen before `compress_projected`, so every
non-decode consumer sees bit-identical fp32 values. Also route the two
compression-off layers' single SWA current-token write through the same
fused kernel (N_CACHES=1) instead of `index_put`.

DeepSeek-V4-Flash proxy(10L) decode window total_us (comm-excl):
  pre  17221.4 us
  post 16497.3 us (-4.20%, keep-path A removal, 21.0x noise floor 0.2%);
  copy_cast instances 710 -> 470 (-240 kernels removed, the targeted
  producer bf16->fp32 casts; abs -604.8 us), gather_scatter 380 -> 360
  (-20 off-layer index_put), other 1510 -> 1530 (+20 fused stores);
  comm swing +3470 us at flat 210 instances = documented NCCL jitter,
  excluded by the gate.
  Output smoke match=false: clean-vs-v2 and clean-vs-(store-only v1)
  diverge at the identical char 189 and the two edited runs diverge from
  each other -- the documented cross-process FP8-atomic nondeterminism of
  the random-weight proxy (same pattern as landed ideas 0004/0006/0044),
  not a numerics change.

Focused tests (all torch.equal/byte-exact, 943 passed DSV4 custom-ops
sweep): test_deepseek_v4_current_token_store.py 40/40 (1..5 cache
arities incl. new off/quad configs, bf16+fp32 value dtypes, extreme
bf16 widening w/ inf/nan/subnormals at page-boundary offsets, CUDA
graph capture+replay x3); test_deepseek_v4_sparse_update_meta_hoist.py
new test_update_reconstruction_dtype_anchored_on_cache (bf16 vs fp32
rows -> byte-identical mhc, R4+R128);
test_deepseek_v4_compressor_raw_rows.py new source-op + compressed-kv
bf16-vs-fp32-widening invariance (R4+R128).
test_deepseek_v4_modeling.py 15 passed / 8 failed IDENTICAL to clean
HEAD (pre-existing CPU-dispatch breakage of trtllm::cublas_mm).
Full-model real-weight TPOT+accuracy confirmation remains an explicit
Tier-2 step.

+++ tensorrt_llm/_torch/auto_deploy/models/custom/modeling_deepseek_v4.py
@@ -1074,8 +1074,13 @@ class DeepseekV4Compressor(nn.Module):
         self.norm = DeepseekV4RMSNorm(self.head_dim, config.rms_norm_eps)
 
     def project(self, hidden_states: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
-        kv_all = _linear_module(hidden_states, self.wkv, layer_type="mla").float()
-        score_all = _linear_module(hidden_states, self.wgate, layer_type="mla").float()
+        # Raw activation-dtype rows (idea_0092): the sparse-attention op widens them
+        # where fp32 math needs it -- the decode current-token store converts in-kernel
+        # (bf16 -> fp32 is exact), so no per-layer ``.float()`` copy runs at decode.
+        # Consumers that need the fp32 tensors up front (``compress_projected`` on the
+        # standalone paths, the op's prefill branch) widen explicitly.
+        kv_all = _linear_module(hidden_states, self.wkv, layer_type="mla")
+        score_all = _linear_module(hidden_states, self.wgate, layer_type="mla")
         return kv_all, score_all
 
     def compress_projected(
@@ -1147,9 +1152,12 @@ class DeepseekV4Compressor(nn.Module):
         position_ids: torch.Tensor,
     ) -> torch.Tensor:
         kv_all, score_all = self.project(hidden_states)
+        # ``compress_projected`` math (pool, ape add) ran on fp32 inputs before
+        # ``project`` started returning raw activation-dtype rows (idea_0092); widen
+        # here so this standalone path stays bit-identical.
         return self.compress_projected(
-            kv_all,
-            score_all,
+            kv_all.float(),
+            score_all.float(),
             cos_table,
             sin_table,
             position_ids,
@@ -1284,9 +1292,12 @@ class DeepseekV4Indexer(nn.Module):
         offset: int,
     ) -> torch.Tensor:
         q, weights, compressor_kv, compressor_gate = self.project(hidden_states, q_lora, cos, sin)
+        # Same widening note as ``DeepseekV4Compressor.forward``: ``project`` returns
+        # raw activation-dtype rows (idea_0092); the standalone indexer path widens
+        # before ``compress_projected`` to keep its math bit-identical.
         index_k = self.compressor.compress_projected(
-            compressor_kv,
-            compressor_gate,
+            compressor_kv.float(),
+            compressor_gate.float(),
             cos_table,
             sin_table,
             position_ids,
